Categorize Macro Interfaces
This commit extends the categorization to macro interfaces
