+++ Examples/Sources/MacroExamples/Implementation/Member/CustomCodable.swift
@@ -54,3 +54,14 @@ public enum CustomCodable: MemberMacro {
     return [codingKeys]
   }
 }
+
+public struct CodableKey: PeerMacro {
+  public static func expansion(
+    of node: AttributeSyntax,
+    providingPeersOf declaration: some DeclSyntaxProtocol,
+    in context: some MacroExpansionContext
+  ) throws -> [DeclSyntax] {
+    // Does nothing, used only to decorate members with data
+    return []
+  }
+}
